+++ src/imio/smartweb/core/contents/pages/page/content.py
@@ -3,6 +3,8 @@
 from imio.smartweb.core.contents import IPages
 from imio.smartweb.core.contents import Pages
 from imio.smartweb.locales import SmartwebMessageFactory as _
+from plone.autoform import directives
+from plone.supermodel import model
 from zope import schema
 from zope.interface import implementer
 
@@ -12,6 +14,8 @@ CATEGORY_TAXONOMY = "collective.taxonomy.page"
 class IPage(IPages):
     """Marker interface and Dexterity Python Schema for Page"""
 
+    directives.order_before(category="ICategorization.subjects")
+    model.fieldset("categorization", fields=["category"])
     category = schema.Choice(
         title=_(u"Category"),
         source=CATEGORY_TAXONOMY,